TransportInfo size mismatches
Summary:
The code that uses TransportInfo assumes these byte counts are
compatible with size_t so make them so.

Test Plan: compiled

Reviewed By: davejwatson@fb.com

Subscribers: fugalh, bmatheny, njormrod, folly-diffs@, benyluo, ranjeeth, fma, kmdent

FB internal diff: D1725026

Tasks: 5538843

Signature: t1:1725026:1418061933:b4072ca926e792716776ddc00d79db9376510859

+++ folly/experimental/wangle/acceptor/TransportInfo.h
@@ -27,13 +27,13 @@ struct HTTPHeaderSize {
    * before decompression. If header compression is not supported, the value
    * is set to 0.
    */
-  uint32_t compressed{0};
+  size_t compressed{0};
 
   /**
    * The number of bytes used to represent the serialized header before
    * compression or after decompression, in plain-text format.
    */
-  uint32_t uncompressed{0};
+  size_t uncompressed{0};
 };
 
 struct TransportInfo {
